pyproject.toml
[project]
name = "ai-simulator"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.3",
    "httpx>=0.27.0",
    "langchain>=0.3.25",
    "langchain-deepseek>=0.1.3",
    "langchain-openai>=0.3.18",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.1",
    "sse-starlette>=3.3.4",
    "uvicorn>=0.43.0",
    "deepagents>=0.4.12",
]
